% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{list_enrollments}
\alias{list_enrollments}
\title{List own round enrollments}
\usage{
list_enrollments(repo, actor, study_id)
}
\arguments{
\item{repo}{Repository.}

\item{actor}{Panel actor.}

\item{study_id}{Selected study.}
}
\value{
Enrollment table.
}
\description{
List own round enrollments
}
