% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{list_studies}
\alias{list_studies}
\title{List studies available to the current actor}
\usage{
list_studies(repo, actor)
}
\arguments{
\item{repo}{Repository.}

\item{actor}{Server actor.}
}
\value{
Study table without contacts or response data.
}
\description{
List studies available to the current actor
}
