% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{new_protocol}
\alias{new_protocol}
\title{Construct a protocol}
\usage{
new_protocol(config)
}
\arguments{
\item{config}{Complete declarative configuration.}
}
\value{
delphyr_protocol; invalid inputs raise DEL_VALIDATION.
}
\description{
Construct a protocol
}
\examples{
new_protocol(demo_protocol())$schema_version
}
