% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{record_worker_heartbeat}
\alias{record_worker_heartbeat}
\title{Record that a worker process is alive}
\usage{
record_worker_heartbeat(repo, worker_id, started = FALSE)
}
\arguments{
\item{repo}{Trusted worker repository.}

\item{worker_id}{Stable name of the worker slot, for example "worker-1".}

\item{started}{TRUE when the process has just started.}
}
\value{
Invisibly TRUE.
}
\description{
Record that a worker process is alive
}
